+++ src/funds/funds_user/router.py
@@ -3,7 +3,7 @@ from typing import List, Dict
 import weaviate
 from fastapi import APIRouter
 
-from src.models import ArticleGet, FundGet
+from src.funds.models import FundGet, FundAdd
 
 router = APIRouter(
     prefix="/funds",
