+++ src/knowledge_base/models.py
@@ -1,0 +1,16 @@
+from typing import List
+
+from pydantic import BaseModel
+
+
+class ArticleAdd(BaseModel):
+    tags: List[str]
+    title: str
+    text: str
+
+
+class ArticleGet(BaseModel):
+    id: str
+    tags: List[str]
+    title: str
+    text: str
